refactor: split the encoder, the dashboard's character layer and the MCP reads
Three files were doing several jobs each. None of the splits change behaviour;
the fixes bundled with them do.

packages/devtools/src/encode.ts held four separable subsystems behind one
960-line module — half of which was prose. It is now the composing entry over
policy.ts (the disclosure clamp), descriptors.ts (the marker vocabulary),
cause.ts (failures as cause chains), shape.ts and full.ts. The cause <-> full
cycle is broken by inverting the seam that already existed: FullContext carries
an injected encodeValue instead of the depth and budget the caller already
holds, which makes cause.ts a leaf.

Fixed along the way:
- an orphaned doc block that had drifted off encodeRecord onto ValueEncoder
- errorFields declared `unknown | undefined`, which is just `unknown`
- ShapeDescriptor was missing its two "full" arms, so the map and set
  descriptors did not typecheck against the union they belong to

apps/hub/src/ui/format.ts takes the nine pure helpers out of App.tsx, with the
contract stated where it can be asserted: the result is never wider than
`width`. Two real bugs there:
- pad guarded on `>=`, so a value that fit its column exactly was ellipsized,
  spending two characters to announce a truncation that had not happened
- neither pad nor clip handled a width too narrow for their marker, so both
  could return more characters than they were given. clip's zero-width case is
  reachable from any terminal under about sixty columns

apps/hub/src/mcpReads.ts takes the resolution layer out of mcp.ts, which leaves
mcp.ts as the wire contract plus the layer and makes the rules testable without
an HTTP server behind them. frond_read_work's conditional `since` spread is
dropped — ReadOptions.since is `number | undefined`, so it was ceremony that
made two adjacent handlers look like they disagreed.

New tests: apps/hub/test/format.test.ts and apps/hub/test/mcp.test.ts. The
second covers the subtlest rule in the hub — the hub's own attachment is
excluded from the sole-candidate fallback but still resolves when named — plus
the fixture detail that makes it possible to get wrong, which is that the hub
and the app it watches both report runtimeId "runtime-1".
